Clarify ${params}
This is tricky because we only want this note to show if they are reading on Github and not in the Extensions > How this Extension Works tab, but thought I would suggest it. 

I didn't realize these parameters would fill in with my specific project details when I saw them on Github.