Update comment for one-time payment
This comments refers to a subcollection on subscription doc, but I think it should ref to the customer doc.

We aren't writing to subscription, but to customer. 

This came up because I'm not using subscriptions but do want to use this to sync payments on one-time purchases (via checkout_session and/or manual invoices)

Overall, it seems like this extension assumes invoice == subscription. Which is fine. I'll just handle certain webhook events outside of this extension. But this is an area where it seems like we can decouple things nicely (FWIW)