Revert "Fix test import statements"
This reverts commit 25300aad9f4b08843bb5b2dc5bd22942c05d47a9.

The previous commit changed import statements for firebase-admin
and firebase-functions to default imports, which caused the extension
deployment to fail with runtime errors (Cannot read properties of
undefined (reading 'auth')). The correct usage is to use namespace
imports (import * as admin from 'firebase-admin'). This revert
restores the working import style and fixes the deployment issue.